+++ windows/flutter/generated_plugins.cmake
@@ -3,7 +3,10 @@
 #
 
 list(APPEND FLUTTER_PLUGIN_LIST
+  app_links
+  cloud_firestore
   connectivity_plus
+  firebase_auth
   firebase_core
 )